requirements.txt
torch==1.12.1+cu116
torchvision==0.13.1+cu116
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
lpips
gdown
pandas
einops
flow_vis
face_alignment
ffmpeg-python==0.2.0
imageio-ffmpeg==0.4.7
timm==0.4.5
py-feat
facenet-pytorch==2.5.3
insightface==0.7.3
onnxruntime-gpu
mediapipe
thop
decord